pyproject.toml
[project]
name = "libefiling"
version = "0.1.17"
authors = [{ name = "hyperion13th144m", email = "hyperion13th144m@gmail.com" }]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asn1crypto (>=1.5.1,<2.0.0)",
    "pillow (>=12.0.0,<13.0.0)",
    "pytesseract (>=0.3.13,<0.4.0)",
    "requests (>=2.32.5,<3.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
]
license = "MIT"
description = "A Python library for e-filing systems."
keywords = ["example", "testpypi", "demo"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
]

[tool.poetry]
packages = [{ include = "libefiling", from = "src" }]
homepage = "https://github.com/hyperion13th144m/libefiling"
repository = "https://github.com/hyperion13th144m/libefiling"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
